int convertion to double

+++ multiplication_template.c
@@ -1,13 +1,13 @@
 #include <stdio.h>
 #include <stdlib.h>
 int main() {
-    int scalarMultiplication(int num1,int num2){
+    double scalarMultiplication(double num1,double num2){
         return num1*num2;
     }
 
-    int *scalarMatrixMultiplication(int scalar, int *matrix, int row, int column){
-        int* newMatrix;
-        newMatrix=malloc(sizeof(int*)*row*column);
+    double *scalarMatrixMultiplication(double scalar, double *matrix, int row, int column){
+        double* newMatrix;
+        newMatrix=malloc(sizeof(double*)*row*column);
         for(int i=0;i<row;i++){
             for(int j=0;j<column;j++){
                 *((newMatrix+i*column) + j) = *((matrix+i*column) + j)*scalar;
@@ -15,20 +15,20 @@ int main() {
         }
         return newMatrix;
     }
-    void printMatrix(int *matrix,int row,int column){
+    void printMatrix(double *matrix,int row,int column){
         for(int i=0;i<row*column;i++){
-            printf("%d\n",*(matrix+i));
+            printf("%f\n",*(matrix+i));
         }
     }
     int row=2;
     int column=2;
 
-    int array[2][2]={2,1,3,5};
+    double array[2][2]={2,1,3,5};
 
-    int *newmatrix=malloc(sizeof(int*)*row*column);
+    double *newmatrix=malloc(sizeof(double*)*row*column);
     newmatrix=scalarMatrixMultiplication(3,(&array)[0][0],row,column);
 
     printMatrix(newmatrix,row,column);
-    
+
    return 0;
 }
